clean build and deploy plan
while hacking on the code the Makefile spins up a docker image
whith postgres and a host mounted volume with the latest app binary.

When deploying run 'fab deploy' to package binary into docker image
and upload it to docker index.
Then on the remote the new image gets pulled and run.